Added HTML comments signifying which script is used for the test
Added HTML comments signifying which specific script tag set is used
for the actual JavaScript for the test. With the comments as they were,
I thought all of the script at the bottom was for the web analytics,
and it made me rather confused until I looked deeper into it.