General HTML improvements (#43)
* Add links to GitHub page

Added links to the GitHub page for the project for the convenience of
the user.

* Changed link names and added links for the email address

Changed link names and added links for the email address on both the
index (front) page and the results page.

* Added HTML comments signifying which script is used for the test

Added HTML comments signifying which specific script tag set is used
for the actual JavaScript for the test. With the comments as they were,
I thought all of the script at the bottom was for the web analytics,
and it made me rather confused until I looked deeper into it.

* Added Dynamic Number of Questions to the Front Page

I added a number that dynamically shows the number of questions
currently in the test.

* Added semicolon at end of JavaScript code

Added semicolon at end of JavaScript code, although it wasn’t
technically needed.