Add some whitespace to the CSS
Greetings! We felt the lines on the homepage were a little close together, so we expanded the line-height.

After that we got excited and added some other margins between elements. We hope you like them!

We only had time to look at the front page today, but hopefully that these changes will also work well on the other pages.