Update questions.js
I rewrote some questions to be more clear.

I replaced "Physician-assisted suicide should be legal" (because it is too specific) with a question about affirmative actions. It is "govt": -10 because it gives privileges without the consent of those who disagree.

Some values were adjusted.
Climate change isn't about dealing with tradition. It is dealing with economy.
Gun ownership should be 10 instead of -10, because it is a "liberty". The right to murder (abortion) is a negative liberty, then I will keep 10 because it is the lack of state intervention.